% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{adhd}
\alias{adhd}
\title{ADHD Synthetic Dataset}
\format{
A data frame with 150 rows and 13 columns
\describe{
\item{ID}{numeric id}
\item{odd}{ODD diagnosis, reflecting whether the child was (coded as 1) or was not (coded as 0)
diagnosed with ODD (oppositional defiant disorder) before the first-stage intervention.}
\item{severity}{ADHD score, reflecting ADHD symptoms at
the end of the previous school year (larger values reflect greater symptoms). Range 0-10.
\item{priormed}{medication prior to first-stage intervention, reflecting whether the child did
(coded as 1) or did not (coded as 0) receive medication during the previous school year.}
\item{race}{race variable: white (coded 1) versus non-white (coded 0).}
\item{Y0}{baseline school performance (higher values reflect better performance).}
\item{A1}{stage 1 treatment assignment. A1 = -1 if the first stage intervention was Medication;
A2 = 1 if the first stage intervention was Behavioral Intervention.}
\item{R}{response status. R = 0 if participant did not respond to the first stage intervention; R= 1 if he or she responded.}
\item{NRtime}{number of months until non-response. Range 2-8.}
\item{adherence}{adherence to the stage 1 intervention: whether the child did (coded as 1) or did not (coded as 0) show high adherence to initial treatment.}
\item{Y1}{school performance at the end of stage 1.}
\item{A2}{stage 2 treatment assignment. Non-responders (R=0) to the first stage intervention entered the second stage of intervention and were re-randomized to either ADDING the other type of intervention (coded as -1) or INTENSIFYING the first stage intervention (coded as 1). For responders A2 = NA.}
\item{Y2}{end-of study school performance}
}
}
\source{
Pelham, et. al.
}
\usage{
adhd
}
\description{
Dataset created for CATIE 2023
}
\section{Operating Characteristics}{
All baseline and moderator variable are grand mean centered. The nuisance terms \code{R.resid} and \code{adherence.resid} are residualized using the true probabilities.
\subsection{Default Structural Nested Mean Model:}{
\subsection{School performance after first-stage}{

\deqn{Y_{1,i}(a_1, a_2) = \eta_0 - \beta_1a_1 + \eta_2U_i + \eta_3Y_{0c}\epsilon_i}
}

\subsection{School performance after second-stage}{

\deqn{Y_{2,i}(a_1, a_2) = \eta_0 - \eta_1odd_c - \eta_2severity_c + \eta_3priormed_c + \eta_4race_c + \\ (\beta_1 - \beta_2priormed_c)a_1 + \\ \eta_5\tilde{R} + \eta_6 \tilde{Adh} + \eta_7\tilde{Y_1} \\ +(1-R)(\beta_3 + \beta_4a_1 + \beta_5Adh_c)a_2 + 0.4U + \epsilon_i}
}

}

\subsection{Baseline Covariates:}{
\itemize{
\item \code{odd}: binary, centered
\item \code{severity}: truncated normal, scaled between -5 to 5
\item \code{priormed}: binary, centered
\item \code{race}: binary, centered
\item \code{U}: an unknown, common cause of \code{R}, \code{adherence}, \code{Y1}, and \code{Y2}. Induces collider bias if naively conditioning on the time-varying covariates.
}
}

\subsection{First-Stage:}{
\itemize{
\item Marginally, \code{BMOD(1)} is initially worse compared to \code{MED(-1)} after the first-stage \eqn{Y_1}, but better in the long run \eqn{Y_2}
\item IF on \code{priormed}, starting with \code{MED} is better
}
}

\subsection{Nuisance Associations}{
\itemize{
\item \code{R}: positively associated with outcome
\item \code{adherence}: positively associated with outcome
\item \code{NRtime}: or time to non-response has no effect
}
}

\subsection{Second-Stage (among non-responders only):}{
\itemize{
\item Marginally, \code{AUG(-1)} is better compared to \code{INT(1)}
\item Positive interaction for \code{AUG} if given \code{MED}, and \code{INT} if given \code{BMOD}
\item IF adherent to first-stage, much better to \code{INT}; if non-adherent to first-stage, much better to \code{AUG}
}
}
}

\keyword{datasets}
