% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{autism}
\alias{autism}
\title{Autism Synthetic Dataset}
\format{
A data frame with 200 rows and 11 columns

#' \describe{
\item{ID}{numeric id}
}
}
\usage{
autism
}
\description{
Dataset created for CATIE 2023
}
\keyword{datasets}
