% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATIE2023-package.R
\docType{package}
\name{CATIE2023-package}
\alias{CATIE2023}
\alias{CATIE2023-package}
\title{CATIE2023: Generate Synthetic SMART Data for CATIE 2023}
\description{
Generates sythetic ADHD and AUTISM SMART data with desired operating characteristics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mferlic/simCATIEdata}
  \item Report bugs at \url{https://github.com/mferlic/simCATIEdata/issues}
}

}
\author{
\strong{Maintainer}: Mason Ferlic \email{mferlic@umich.edu}

}
\keyword{internal}
