% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{estimate}
\usage{
estimate(fit, combos)
}
\arguments{
\item{fit}{model object}

\item{combos}{vector of linear combinations of parameter estimates.}
}
\value{
prints Coefmat
}
\description{
Defines a rough equivalent to proc genmod's estimate statement.
}
