% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{print.estimate}
\alias{print.estimate}
\title{print.estimate}
\usage{
\method{print}{estimate}(object, digits = 4, signif.stars = TRUE, ...)
}
\arguments{
\item{object}{of class "estimate"}

\item{digits}{number of sig digits to display}

\item{signif.stars}{show stars?}

\item{...}{}
}
\value{
printCoefmat
}
\description{
print.estimate
}
