% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simADHDsmart.R
\name{simADHDsmart}
\alias{simADHDsmart}
\title{simADHDsmart}
\usage{
simADHDsmart(
  N = 150,
  baseline.params = list(p.odd = 0.4, m.severity = 0, p.priormed = 0.3, p.race = 0.8),
  Y0.coef = c(2, odd = -0.2, severity = -0.3),
  U.params = c(mu = 0, sd = 1),
  R.coef = c(-0.4, A1 = -0.1, `priormed:A1` = -0.2, U = 0.2),
  adherence.coef = c(-0.1, `priormed:A1` = -0.2, U = 0.1),
  event_time.coef = NULL,
  Y1.coef = c(2.5, A1 = -0.3, U = 0.1),
  Y2.baseline = c(3, odd = -0.3, severity = -0.4, priormed = 0, race = 0.5),
  Y2.tx1 = c(A1 = 0.3, `priormed:A1` = -1.4),
  Y2.n1 = c(R.resid = 0.8, adherence.resid = 1, U = 0.4),
  Y2.tx2 = c(A2 = -0.3, `A1:A2` = 0.1, `adherence:A2` = 1.2),
  sigma = 1
)
}
\arguments{
\item{N}{number of observations to generate}

\item{baseline.params}{A list specifying the probability \code{odd}, mean \code{severity}, prob. \code{priormed}, prob. \code{race}}

\item{Y0.coef}{A named vector specifying the linear model coefficients for baseline school performance. Can be a function of any past variables. See \emph{Named Vectors}}

\item{U.params}{A vector specifying the normal distribution parameters \code{mu} and \code{sd}}

\item{R.coef}{A named vector specifying the linear \strong{probit model} coefficients for the probability of being a responder to first-stage treatment. Can be a function of any past variables.}

\item{adherence.coef}{A named vector specifying the linear \strong{probit model} coefficients for the probability of being adherent to first-stage treatment. Can be a function of any past variables.}

\item{event_time.coef}{not implemented}

\item{Y1.coef}{A named vector of linear model coefficients specifying the first-stage treatment causal effect on \eqn{Y_1} school performance}

\item{Y2.baseline}{A named vector of linear model coefficients specifying the baseline covariate associations on end-of-study \eqn{Y_2} school performance}

\item{Y2.tx1}{A named vector of linear model coefficients specifying the first-stage treatment causal effect on end-of-study \eqn{Y_2} school performance. Can be a function of any baseline moderators.}

\item{Y2.n1}{A named vector of linear model coefficients specifying the nuisance associations on end-of-study \eqn{Y_2} school performance. Can be a function of any prior moderators. NOTE: must specify R.resid and adherence.resid which are the direct associations in the SNMM.}

\item{Y2.tx2}{A named vector of linear model coefficients specifying the second-stage treatment causal effect, among non-responders, on end-of-study \eqn{Y_2} school performance. Can be a function of any baseline moderators. NOTE: all moderators are grand mean centered}

\item{sigma}{gaussian noise added to \eqn{Y_{0,1,2}}}
}
\value{
A list with components
\describe{
\item{data}{data.frame of observed variables}
\item{DTRmean}{marginal embedded DTR means}
}
}
\description{
This function simulates synthetic data to mirror desired characteristics of the Pelham ADHD SMART study.
Created specifically for CATIE 2023.
}
\section{Named Vectors}{
The named vectors of coefficients use formula notation to specify interactions i.e. "\code{priormed:A1}", which must be quoted. The order does not matter but spelling does. Note: if the first element is unamed it is treated as the intercept term, else if all elements are named the intercept is ommited.
}

\section{Operating Characteristics}{
All baseline and moderator variable are grand mean centered. The nuisance terms \code{R.resid} and \code{adherence.resid} are residualized using the true probabilities.
\subsection{Default Structural Nested Mean Model:}{
\subsection{School performance after first-stage}{

\deqn{Y_{1}(a_1, a_2) = 2.5 - 0.3a_1 + 0.4U + \epsilon}
}

\subsection{School performance after second-stage}{

\deqn{Y_{2}(a_1, a_2) = 3 - 0.2odd - 0.3severity + 0*priormed + .5race + \\ (0.3 - 1.4priormed)a_1 + \\ 0.8R.resid(a_1, u) + 1 adherence.resid(a_1, u) \\ +(1-R)(-0.3 + 0.1a_1 + 1.2adherence)a_2 + 0.4U + \epsilon}
}

}

\subsection{Baseline Covariates:}{
\itemize{
\item \code{odd}: binary, centered
\item \code{severity}: standard normal
\item \code{priormed}: binary, centered
\item \code{race}: binary, centered
\item \code{U}: an unknown, common cause of \code{R}, \code{adherence}, \code{Y1}, and \code{Y2}. Induces collider bias if naively conditioning on the time-varying covariates.
}
}

\subsection{First-Stage:}{
\itemize{
\item Marginally, \code{BMOD(1)} is initially worse compared to \code{MED(-1)} after the first-stage \eqn{Y_1}, but better in the long run \eqn{Y_2}
\item IF on \code{priormed}, starting with \code{MED} is better
}
}

\subsection{Nuisance Associations}{
\itemize{
\item \code{R}: positively associated with outcome
\item \code{adherence}: positively associated with outcome
\item \code{event-time}: or time to non-response has no effect
}
}

\subsection{Second-Stage (among non-responders only):}{
\itemize{
\item Marginally, \code{AUG(-1)} is better compared to \code{INT(1)}
\item Positive interaction for \code{AUG} if given \code{MED}, and \code{INT} if given \code{BMOD}
\item IF adherent to first-stage, much better to \code{INT}; if non-adherent to first-stage, much better to \code{AUG}
}
}
}

\section{Internal functions}{
\itemize{
\item linearMult: takes a named vector of coefficients and a data.frame and creates the necessary design matrix to multiply and return the function values.
\item sampleProbitMean: generates observations from the posterior in order to empirically evaluate probability the integral.
\item getMarginalMeans: takes the specified coefficients and returns the marginal means of the four embedded adaptive interventions, averaging over response and adherence.
}
}

\author{
Mason Ferlic, December 2022, University of Michigan, D3C
}
