% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simAUTISMsmart.R
\name{simAUTISMsmart}
\alias{simAUTISMsmart}
\title{simAUTISMsmart}
\usage{
simAUTISMsmart(
  N = 200,
  baseline.params = list(O11 = c(lo = 0, up = 80), O12 = c(2, 4)),
  Y0.coef = c(40, O11 = 0.4, O12 = 0),
  U.params = c(mu = 0, sd = 1),
  R.coef = c(0.25, A1 = -0.3, U = 0.1),
  O21.coef = c(5.6, A1 = -0.4, R = 2.5, U = 5),
  O22.coef = c(48, A1 = 3, R = 8, `A1:R` = -4, O21 = -2, U = 4),
  Y1.coef = c(50, A1 = 1, U = 4),
  Y2.baseline = c(60, O11 = 0.3),
  Y2.tx1 = c(A1 = 1, `A1:O11` = 0.3),
  Y2.n1 = c(R.resid = 18, O21.resid = 5, U = 6),
  Y2.tx2 = c(A2 = -4, `A2:O21` = 6),
  sigma = 8
)
}
\arguments{
\item{N}{number of observations to generate.}

\item{baseline.params}{A named list specifying the \code{unif} parameters for O11 and \code{beta} parameters for O12.}

\item{Y0.coef}{A named vector specifying the linear model coefficients for score. Can be a function of any past variables. See \emph{Named Vectors}.}

\item{U.params}{A vector specifying the normal distribution parameters \code{mu} and \code{sd}.}

\item{R.coef}{A named vector specifying the \strong{linear probit model} coefficients for the probability of being a responder to first-stage treatment. Can be a function of any past variables.}

\item{O21.coef}{A named vector specifying the \strong{linear model} coefficients for O21. Can be a function of any past variables.}

\item{O22.coef}{A named vector specifying the \strong{linear model} coefficients for O22. Can be a function of any past variables.}

\item{Y1.coef}{A named vector of \strong{linear model} coefficients specifying the first-stage treatment causal effect on \eqn{Y_1}.}

\item{Y2.baseline}{A named vector of \strong{linear model} coefficients specifying the baseline covariate associations on end-of-study \eqn{Y_2}.}

\item{Y2.tx1}{A named vector of \strong{linear model} coefficients specifying the first-stage treatment causal effect on end-of-study \eqn{Y_2} school performance. Can be a function of any baseline moderators.}

\item{Y2.n1}{A named vector of \strong{linear model} coefficients specifying the nuisance associations on end-of-study \eqn{Y_2}. Can be a function of any prior moderators. NOTE: must specify O21.resid and O22.resid which are the residualized direct associations in the SNMM.}

\item{Y2.tx2}{A named vector of \strong{linear model} coefficients specifying the second-stage treatment causal effect, among non-responders to a1 == 1, on end-of-study \eqn{Y_2}. Can be a function of any baseline moderators. NOTE: all moderators are residualized centered}

\item{sigma}{Gaussian noise added to \eqn{Y_{0,1,2}}}
}
\value{
A data.frame with attributes
\describe{
\item{simparams}{list of DGM coeffecients used in function call}
\item{dtrmeans}{marginal embedded DTR means}
\item{creator}{System USER who generated the data}
\item{time_created}{time of data realization}
}
}
\description{
This function simulates synthetic data to roughly mirror desired characteristics of the AUTISM SMART study by Kasari.
Created specifically for CATIE 2023.
}
\section{Named Vectors}{
The named vectors of coefficients use formula notation to specify interactions i.e. "\code{A1:O11}", which must be quoted. The order does not matter but spelling does. Note: if the first element is unamed it is treated as the intercept term, else if all elements are named the intercept is omitted.
}

\section{Operating Characteristics}{
All baseline variables are grand mean centered. All moderators are centered conditional on the expectation given the past (residuals).
\subsection{Default Structural Nested Mean Model:}{
\subsection{Y1 after first-stage}{

\deqn{Y_{1,i}(a_1, a_2) = 50 + 1a_1 + 4U_i + \epsilon_i}
}

\subsection{Y2 after second-stage}{

\deqn{Y_{2,i}(a_1, a_2) = 60 + 0.3\bar{O}_{11} + (1 + 0.3\bar{O}_{11})a_1 + \\ 18\tilde{R} + 5\tilde{O}_{21} \\ +I(a_1 =1)(1-R)(-4 + 6\tilde{O}_{21})a_2 + 6U_i + \epsilon_i}
}

}
}

\section{Internal functions}{
\itemize{
\item linearMult: takes a named vector of coefficients and a data.frame and creates the necessary design matrix to multiply and return the function values.
\item sampleProbitMean: generates observations from the posterior in order to empirically evaluate probability the integral.
\item getMarginalMeans: takes the specified coefficients and returns the marginal means of the four embedded adaptive interventions, averaging over response and adherence.
}
}

\author{
Mason Ferlic, December 2022, University of Michigan, D3C
}
