% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATIE2023-package.R
\docType{package}
\name{CATIE2023-package}
\alias{CATIE2023}
\alias{CATIE2023-package}
\title{CATIE2023: Generate Synthetic SMART Data for CATIE 2023}
\description{
Contains the synthetic ADHD and AUTISM SMART data sets with the desired operating characteristics for use with CATIE 2023. Also contains the data generating code along with any helper fucntions for the demo days.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mferlic/simCATIEdata}
  \item Report bugs at \url{https://github.com/mferlic/simCATIEdata/issues}
}

}
\author{
\strong{Maintainer}: Mason Ferlic \email{mferlic@umich.edu}

}
\keyword{internal}
